Move strong_typedef.h into core
