+++ include/ozo/type_traits.h
@@ -2,7 +2,9 @@
 
 #include <ozo/detail/pg_type.h>
 #include <ozo/detail/float.h>
-#include <ozo/detail/strong_typedef.h>
+#include <ozo/core/strong_typedef.h>
+#include <ozo/core/nullable.h>
+#include <ozo/core/unwrap.h>
 
 #include <libpq-fe.h>
 
@@ -91,224 +93,6 @@ using null_oid_t = oid_constant<0>;
  * @ingroup group-type_system-constants
  */
 constexpr null_oid_t null_oid;
-
-template <typename T, typename Enable = void>
-struct is_nullable : std::false_type {};
-
-/**
- * @brief Indicates if type meets nullable requirements.
- *
- * `Nullable` type has to:
- * * have a null-state,
- * * be `bool` convertable - `false` indicates null state,
- * * be dereferenceable via `operator *`,
- * * have `allocate_nullable()` specialization (see the function documentation for how to).
- *
- * These next types are `Nullable` out of the box:
- * * `boost::optional`,
- * * `std::optional`,
- * * `boost::scoped_ptr`,
- * * `std::unique_ptr`,
- * * `boost::shared_ptr`,
- * * `std::shared_ptr`,
- * * `boost::weak_ptr`,
- * * `std::weak_ptr`.
- *
- * @note Raw pointers are not supported as `Nullable` by default, because this library uses RAII model for objects and
- * no special deallocation functions are used. But if it is really needed and you want to deallocate the allocated
- * objects manually you can add them as described below.
- *
- * ### Example
- *
- * If you want to add nullable type to the library --- you have to specialize `ozo::is_nullable` struct and specify allocation function like this:
- *
- * @code
-//Add raw pointer to nullables
-
-namespace ozo {
-
-// Mark raw pointer as nullable
-template <typename T>
-struct is_nullable<T*> : std::true_type {};
-
-// Specify allocation function
-template <typename T>
-struct allocate_nullable_impl<T*> {
-    template <typename Alloc>
-    static void apply(T*& out, const Alloc&) {
-        out = new T();
-    }
-};
-
-}
-
- * @endcode
- *
- * @sa ozo::unwrap(), is_null(), allocate_nullable(), init_nullable(), reset_nullable()
- * @ingroup group-core-concepts
- * @hideinitializer
- */
-template <typename T>
-constexpr auto Nullable = is_nullable<std::decay_t<T>>::value;
-
-namespace detail {
-
-struct unwrap_with_forward {
-    template <typename T>
-    constexpr static decltype(auto) apply(T&& v) noexcept {
-        return std::forward<T>(v);
-    }
-};
-
-struct unwrap_with_dereference {
-    template <typename T>
-    constexpr static decltype(auto) apply(T&& v) noexcept(noexcept(*v)) {
-        return *v;
-    }
-};
-
-struct unwrap_with_get {
-    template <typename T>
-    constexpr static decltype(auto) apply(T&& v) noexcept(noexcept(v.get())) {
-        return v.get();
-    }
-};
-
-template <typename T, typename = std::void_t<>>
-struct unwrap_impl_default : unwrap_with_forward {};
-
-template <typename T>
-struct unwrap_impl_default <T, Require<Nullable<T>>> : unwrap_with_dereference {};
-
-} // namespace detail
-
-
-template <typename T>
-struct unwrap_impl : detail::unwrap_impl_default<T> {};
-
-/**
- * @brief Dereference #Nullable argument or forward it
- *
- * This utility function gets value from #Nullable, and pass object as it is if it is not #Nullable
- *
- * @param value --- object to unwrap
- * @return dereferenced argument in case of #Nullable, pass as is if it is not.
- * @ingroup group-core-functions
- */
-template <typename T>
-constexpr decltype(auto) unwrap(T&& v) noexcept(
-        noexcept(unwrap_impl<std::decay_t<T>>::apply(std::forward<T>(v)))) {
-    return unwrap_impl<std::decay_t<T>>::apply(std::forward<T>(v));
-}
-
-template <typename T>
-struct unwrap_impl<std::reference_wrapper<T>> : detail::unwrap_with_get {};
-
-namespace detail {
-
-template <typename T, typename V>
-struct is_null_always {
-    constexpr static V apply(const T&) noexcept { return {};}
-};
-
-template <typename T, typename = std::void_t<>>
-struct is_null_default_impl : is_null_always<T, std::false_type> {};
-
-template <typename T>
-struct is_null_default_impl<T, Require<Nullable<T>>> {
-    constexpr static auto apply(const T& v) noexcept(noexcept(!v)) {
-        return !v;
-    }
-};
-
-} // namespace detail
-
-template <typename T>
-struct is_null_impl : detail::is_null_default_impl<T> {};
-
-/**
- * @brief Indicates if value is in null state
- *
- * This utility function indicates if given value is in null state.
- * If argument is not #Nullable it always returns `false`
- *
- * @param value --- object to examine
- * @return `true` --- object is in null-state,
- * @return `false` --- overwise.
- * @ingroup group-core-functions
- */
-template <typename T>
-constexpr auto is_null(const T& v) noexcept(noexcept(is_null_impl<T>::apply(v))) {
-    return is_null_impl<T>::apply(v);
-}
-
-template <typename T, typename = std::void_t<>>
-struct allocate_nullable_impl {
-    static_assert(Emplaceable<T>, "default implementation uses emplace() method");
-    template <typename Alloc>
-    static void apply(T& out, const Alloc&) {
-        out.emplace();
-    }
-};
-
-template <typename T, typename Alloc>
-inline void allocate_nullable(T& out, const Alloc& a) {
-    static_assert(Nullable<T>, "T must be nullable");
-    allocate_nullable_impl<std::decay_t<T>>::apply(out, a);
-}
-
-template <typename T, typename Alloc = std::allocator<char>>
-inline void init_nullable(T& n, const Alloc& a = Alloc{}) {
-    static_assert(Nullable<T>, "T must be nullable");
-    if (is_null(n)) {
-        allocate_nullable(n, a);
-    }
-}
-
-template <typename T>
-inline void reset_nullable(T& n) {
-    static_assert(Nullable<T>, "T must be nullable");
-    n = T{};
-}
-
-/**
- * @brief Unwraps nullable or reference wrapped type
- * @ingroup group-type_system-types
- * Sometimes it is needed to know the underlying type of #Nullable or
- * type is wrapped with `std::reference_type`. So that is why the type exists.
- * Convenient shortcut is `ozo::unwrap_type`.
- * @tparam T -- type to unwrap.
- *
- * ### Example
- *
- * @code
-int a = 42;
-auto ref_a = std::ref(a);
-auto opt_a = std::make_optional(a);
-
-static_assert(std::is_same<ozo::unwrap_type<decltype(ref_a)>, decltype(a)>);
-static_assert(std::is_same<ozo::unwrap_type<decltype(opt_a)>, decltype(a)>);
- * @endcode
- *
- * @sa ozo::unwrap_type, ozo::unwrap
- */
-template <typename T>
-struct get_unwrapped_type {
-#ifdef OZO_DOCUMENTATION
-    using type = <unwrapped type>; //!< Unwrapped type
-#else
-    using type = std::decay_t<decltype(unwrap(std::declval<T>()))>;
-#endif
-};
-
-/**
- * @brief Shortcut for `ozo::unwrap_type`.
- * @ingroup group-type_system-types
- *
- * @tparam T -- type to unwrap.
- */
-template <typename T>
-using unwrap_type = typename get_unwrapped_type<T>::type;
 
 /**
  * @brief Indicates if type is PostgreSQL array representation.
